# api/openapi.runtime.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "JobPilot Platform API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/users/me": {
   "get": {
    "tags": [
     "profile"
    ],
    "summary": "Get Me",
    "operationId": "get_me_api_v1_users_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "patch": {
    "tags": [
     "profile"
    ],
    "summary": "Update Me",
    "operationId": "update_me_api_v1_users_me_patch",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/users/me/preferences": {
   "get": {
    "tags": [
     "profile"
    ],
    "summary": "Get Preferences",
    "operationId": "get_preferences_api_v1_users_me_preferences_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PreferencesIO-Output"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "put": {
    "tags": [
     "profile"
    ],
    "summary": "Put Preferences",
    "operationId": "put_preferences_api_v1_users_me_preferences_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PreferencesIO-Input"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PreferencesIO-Output"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/users/me/notification-settings": {
   "get": {
    "tags": [
     "profile"
    ],
    "summary": "Get Notification Settings",
    "operationId": "get_notification_settings_api_v1_users_me_notification_settings_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NotificationSettingsIO"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "put": {
    "tags": [
     "profile"
    ],
    "summary": "Put Notification Settings",
    "operationId": "put_notification_settings_api_v1_users_me_notification_settings_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NotificationSettingsIO"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NotificationSettingsIO"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LocationPref": {
    "properties": {
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "state": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "State"
     },
     "country": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Country"
     },
     "radius_mi": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Radius Mi"
     }
    },
    "type": "object",
    "title": "LocationPref"
   },
...
   "NotificationSettingsIO": {
    "properties": {
     "email_enabled": {
      "type": "boolean",
      "title": "Email Enabled",
      "default": true
     },
     "daily_report_hour": {
      "type": "integer",
      "maximum": 23.0,
      "minimum": 0.0,
      "title": "Daily Report Hour",
      "default": 21
     },
     "slack_webhook": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Slack Webhook"
     },
     "discord_webhook": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Discord Webhook"
     }
    },
    "type": "object",
    "title": "NotificationSettingsIO",
    "description": "GET returns webhooks masked; PUT treats masked values as 'unchanged'\nand empty strings as 'clear'."
   },
   "PreferencesIO-Input": {
    "properties": {
     "desired_titles": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Desired Titles",
      "default": []
     },
     "employment_types": {
      "items": {
       "type": "string",
       "enum": [
        "FULL_TIME",
        "PART_TIME",
        "CONTRACT",
        "INTERNSHIP",
        "TEMPORARY",
        "UNKNOWN"
       ]
      },
      "type": "array",
      "title": "Employment Types",
      "default": [
       "FULL_TIME"
      ]
     },
     "contract_arrangements": {
      "items": {
       "type": "string",
       "enum": [
        "W2",
        "C1099",
        "C2C",
        "UNSPECIFIED"
       ]
      },
      "type": "array",
      "title": "Contract Arrangements",
      "default": []
     },
     "workplace_types": {
      "items": {
       "type": "string",
       "enum": [
        "REMOTE",
        "HYBRID",
        "ONSITE",
        "UNKNOWN"
       ]
      },
      "type": "array",
      "title": "Workplace Types",
      "default": []
     },
     "locations": {
      "items": {
       "$ref": "#/components/schemas/LocationPref"
      },
      "type": "array",
      "title": "Locations",
      "default": []
     },
     "countries": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Countries",
      "default": [
       "US"
      ]
     },
     "seniority": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "ENTRY",
         "MID",
         "SENIOR",
         "LEAD",
         "PRINCIPAL",
         "UNKNOWN"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Seniority"
     },
     "years_experience": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Years Experience"
     },
     "visa_status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Visa Status"
     },
     "work_authorization": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Work Authorization"
     },
     "needs_sponsorship": {
      "type": "boolean",
      "title": "Needs Sponsorship",
      "default": false
     },
     "open_to_staffing": {
      "type": "boolean",
      "title": "Open To Staffing",
      "default": true
     },
     "salary_min": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salary Min"
     },
     "salary_max": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salary Max"
     },
     "salary_currency": {
      "type": "string",
      "title": "Salary Currency",
      "default": "USD"
     },
     "availability_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Availability Date"
     },
     "notice_period_days": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notice Period Days"
     },
     "auto_apply_enabled": {
      "type": "boolean",
      "title": "Auto Apply Enabled",
      "default": false
     },
     "auto_apply_min_score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       }
      ],
      "title": "Auto Apply Min Score",
      "default": "70"
     },
     "auto_apply_daily_cap": {
      "type": "integer",
      "maximum": 200.0,
      "minimum": 0.0,
      "title": "Auto Apply Daily Cap",
      "default": 25
     }
    },
    "type": "object",
    "title": "PreferencesIO",
    "description": "Single schema for GET response and PUT request (full replace)."
   },
   "PreferencesIO-Output": {
    "properties": {
     "desired_titles": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Desired Titles",
      "default": []
     },
     "employment_types": {
      "items": {
       "type": "string",
       "enum": [
        "FULL_TIME",
        "PART_TIME",
        "CONTRACT",
        "INTERNSHIP",
        "TEMPORARY",
        "UNKNOWN"
       ]
      },
      "type": "array",
      "title": "Employment Types",
      "default": [
       "FULL_TIME"
      ]
     },
     "contract_arrangements": {
      "items": {
       "type": "string",
       "enum": [
        "W2",
        "C1099",
        "C2C",
        "UNSPECIFIED"
       ]
      },
      "type": "array",
      "title": "Contract Arrangements",
      "default": []
     },
     "workplace_types": {
      "items": {
       "type": "string",
       "enum": [
        "REMOTE",
        "HYBRID",
        "ONSITE",
        "UNKNOWN"
       ]
      },
      "type": "array",
      "title": "Workplace Types",
      "default": []
     },
     "locations": {
      "items": {
       "$ref": "#/components/schemas/LocationPref"
      },
      "type": "array",
      "title": "Locations",
      "default": []
     },
     "countries": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Countries",
      "default": [
       "US"
      ]
     },
     "seniority": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "ENTRY",
         "MID",
         "SENIOR",
         "LEAD",
         "PRINCIPAL",
         "UNKNOWN"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Seniority"
     },
     "years_experience": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Years Experience"
     },
     "visa_status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Visa Status"
     },
     "work_authorization": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Work Authorization"
     },
     "needs_sponsorship": {
      "type": "boolean",
      "title": "Needs Sponsorship",
      "default": false
     },
     "open_to_staffing": {
      "type": "boolean",
      "title": "Open To Staffing",
      "default": true
     },
     "salary_min": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salary Min"
     },
     "salary_max": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salary Max"
     },
     "salary_currency": {
      "type": "string",
      "title": "Salary Currency",
      "default": "USD"
     },
     "availability_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Availability Date"
     },
     "notice_period_days": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notice Period Days"
     },
     "auto_apply_enabled": {
      "type": "boolean",
      "title": "Auto Apply Enabled",
      "default": false
     },
     "auto_apply_min_score": {
      "type": "string",
      "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
      "title": "Auto Apply Min Score",
      "default": "70"
     },
     "auto_apply_daily_cap": {
      "type": "integer",
      "maximum": 200.0,
      "minimum": 0.0,
      "title": "Auto Apply Daily Cap",
      "default": 25
     }
    },
    "type": "object",
    "title": "PreferencesIO",
    "description": "Single schema for GET response and PUT request (full replace)."
   },
...
   "UserOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "full_name": {
      "type": "string",
      "title": "Full Name"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "timezone": {
      "type": "string",
      "title": "Timezone"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "full_name",
     "role",
     "timezone",
     "created_at"
    ],
    "title": "UserOut"
   },
   "UserPatch": {
    "properties": {
     "full_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Full Name"
     },
     "timezone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Timezone"
     }
    },
    "type": "object",
    "title": "UserPatch"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}